fix arithmetic shift right

+++ include/misc.h
@@ -24,7 +24,7 @@ static inline int32_t sign_extend(uint32_t value, int bits)
 
 static inline uint32_t shift_right_arith(uint32_t value, int bits)
 {
-    return ~((~value) >> (bits));
+    return ~(((int32_t)(~value)) >> (bits));
 }
 
 #endif
